243: move GetBlockByNumberParams
