+++ src/rpc_client/requests.rs
@@ -142,3 +142,19 @@ impl Display for BlockTag {
         }
     }
 }
+
+/// Parameters of the [`eth_getBlockByNumber`](https://ethereum.org/en/developers/docs/apis/json-rpc/#eth_getblockbynumber) call.
+#[derive(Debug, Serialize, Clone)]
+#[serde(into = "(BlockSpec, bool)")]
+pub struct GetBlockByNumberParams {
+    /// Integer block number, or "latest" for the last mined block or "pending", "earliest" for not yet mined transactions.
+    pub block: BlockSpec,
+    /// If true, returns the full transaction objects. If false, returns only the hashes of the transactions.
+    pub include_full_transactions: bool,
+}
+
+impl From<GetBlockByNumberParams> for (BlockSpec, bool) {
+    fn from(value: GetBlockByNumberParams) -> Self {
+        (value.block, value.include_full_transactions)
+    }
+}
